Proposal: optionally allow auto-translate
The current PR raises an exception if you try to directly map
from one Vocabulary to another.  This is definitely the direction
we want to go in, to force people to be explicit about the translation
by making Actions like ```--> state2=translate(state1)```.  However, rather
than breaking any existing code that makes use of this auto-translation,
we can instead give a DeprecationWarning and fill in the
```translate()``` automatically.  However, if people want the strict
behaviour, they can turn that on at the Module level by indicating
```module.strict=True```.

I would propose that this version (that defaults to the current master
behaviour but allows strict mode if requested) be merged into the nengo
master, but that the eventual nengo_spa repository would only have the
strict behaviour.